==> 2)grocery_products_scraper.ipynb <==
import random
import time
from selenium_stealth import stealth
from selenium import webdriver
# --- 1. User-Agent Setup ---
USER_AGENTS = [
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/118.0.0.0 Safari/537.36",
    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/118.0.0.0 Safari/537.36",
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/117.0.0.0 Safari/537.36",
]
random_user_agent = random.choice(USER_AGENTS)

# --- 2. Proxy Setup ---
# Note: Replace USERNAME and PASSWORD with your actual credentials
#proxy_options = {
#    'proxy': {
#        'no_proxy': 'localhost,127.0.0.1' # Bypasses proxy for local traffic
#    }
#}

# --- 3. Chrome Options Setup ---
chrome_options = webdriver.ChromeOptions()
chrome_options.add_argument(f"--user-agent={random_user_agent}")
chrome_options.add_argument("--start-maximized")
# chrome_options.add_argument("--headless") # Commented out to see the browser
chrome_options.add_argument("--lang=en-US,en;q=0.9")

# Basic anti-detection options
chrome_options.add_experimental_option("excludeSwitches", ["enable-automation"])
chrome_options.add_experimental_option('useAutomationExtension', False)

# Load only DOM content
chrome_options.page_load_strategy = 'eager' 

# Initialize driver with selenium-wire
driver = webdriver.Chrome(
    options=chrome_options,
    #seleniumwire_options=proxy_options
)






# --- 4. Stealth Setup ---
stealth(driver,
        languages=["en-US", "en"],
        vendor="Google Inc.",
        platform="Win32",
        webgl_vendor="Intel Inc.",
        renderer="Intel Iris OpenGL Engine",
        fix_hairline=True,
        )

# --- 5. Run the Scraper ---
print("Scraper is running with a stealth configuration!")
print(f"Using IP from proxy and User-Agent: {random_user_agent}")


from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import TimeoutException, NoSuchElementException
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.action_chains import ActionChains

from urllib.parse import urlparse, parse_qs, unquote
 
 
import json
import time
import random
from pprint import pprint

import pandas as pd

#def infinite_scroll(driver):
    #infinte scroll to load all products
    #last_height = driver.execute_script("return document.body.scrollHeight")
    #while True:
       # driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
        
        #time.sleep(6)
        
       # new_height = driver.execute_script("return document.body.scrollHeight")
        #if new_height == last_height:
         #   break
       # last_height = new_height
       
def infinite_scroll(driver):
    # Find the body element to send keys to
    body = driver.find_element(By.TAG_NAME, 'body')
    
    last_height = driver.execute_script("return document.body.scrollHeight")
    
    while True:
        # Scroll down by one "page"
        for i in range(5):
            body.send_keys(Keys.PAGE_DOWN)
            time.sleep(2)  # Random sleep to mimic human behavior
        
        new_height = driver.execute_script("return document.body.scrollHeight")
        if new_height == last_height:
            print("Reached the end of the page.")
            break
        last_height = new_height

for alfateh

def scraope_al_fateh_grocery_products(driver):
    all_products = []
    STORE_NAME="al-fateh"
    AL_FATEH_GROCERY_URL = "https://alfatah.pk/pages/grocery-foods"


    try:
        #get to grocery foods page
        driver.get(AL_FATEH_GROCERY_URL)
        
        #wait for all categories to load
        wait=WebDriverWait(driver, 10)
        category_boxes=wait.until(
            EC.presence_of_all_elements_located((By.CLASS_NAME, "box"))
        )
        print("Al-Fateh Grocery page loaded successfully.")
        category_boxes_links = [box.find_element(By.TAG_NAME, "a").get_attribute("href") for box in category_boxes]
        categories = [box.text for box in category_boxes]
        
        
        #for each category, click and wait for products to load
        for category_link, CATEGORY_NAME in zip(category_boxes_links, categories):
    
            time.sleep(random.uniform(2, 7))
            #click on the category
            driver.get(category_link)
            
            
            wait.until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, ".col-6.col-sm-6.col-md-4.col-lg-2")))
            infinite_scroll(driver)
            products=wait.until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, ".col-6.col-sm-6.col-md-4.col-lg-2")))
            products_count = len(products)
            
            
            for product in products:
                NAME = None
                PRICE = None
                PRODUCT_LINK = None
                
                PRICE=product.find_element(By.CLASS_NAME, "product-price").text.strip()
                NAME=product.find_element(By.CLASS_NAME, "product-title-ellipsis").text.strip()
                PRODUCT_LINK=product.find_element(By.TAG_NAME, "a").get_attribute("href")
                image_container=product.find_element(By.CLASS_NAME, "image")
                image_url=image_container.find_element(By.TAG_NAME, "img").get_attribute("src")
                
                all_products.append({
                    "store": STORE_NAME,
                    "category":CATEGORY_NAME.strip(),
                    "name": NAME,
                    "price": PRICE,
                    "product-link": PRODUCT_LINK,
                    "image_url": image_url
                })
                
            print (f"Category '{CATEGORY_NAME}': Found {len(products)} products")
            #pprint(all_products)
        
    except Exception as e:
        print(f"An error occurred: {e}")
        driver.save_screenshot("error_screenshot.png")

    finally:
        return all_products

for jalalsons


def scrape_jalal_sons_grocery_products(driver):
    all_products = []
    STORE_NAME="jalal-sons"
    JALAL_SONS_URL = "https://jalalsons.com.pk/"

    try:
            store_name = "Jalal Sons"
            JALALSONS_GROCERY_URL = f"https://jalalsons.com.pk/"
            
            driver.get(JALALSONS_GROCERY_URL)
            wait = WebDriverWait(driver, 20)
            
            # Close banner if present
            try:
                banner_close_button = wait.until(EC.element_to_be_clickable(
                    (By.CSS_SELECTOR, ".cursor-pointer.ms-auto")
                ))
                banner_close_button.click()
            except TimeoutException:
                print(f"[{store_name}] No banner appeared")
            
            # Select location from dropdown
            try:
                from selenium.webdriver.support.ui import Select
                location_dropdown = wait.until(EC.presence_of_element_located(
                    (By.ID, "selectDeliveryBranch")
                ))
                select_object = Select(location_dropdown)
                all_options = select_object.options
                
                enabled_options = [
                    opt for opt in all_options
                    if opt.is_enabled() and opt.get_attribute('value') != ""
                ]
                
                if enabled_options:
                    random_option = random.choice(enabled_options)
                    select_object.select_by_visible_text(random_option.text)
                    
                    try:
                        submit_button = driver.find_element(By.CLASS_NAME, "current_loc_pop_btn")
                        submit_button.click()
                    except Exception as e:
                        print(f"[{store_name}] No button to confirm location selection: {str(e)}")
            except:
                print(f"[{store_name}] No location box appeared")
            

            #build a action chain to move to dropdown and perform
            from selenium.webdriver.common.action_chains import ActionChains
            actions=ActionChains(driver)
            location_dropdown=driver.find_element(By.XPATH, "/html/body/header[3]/div/nav/div[1]/ul/li[8]")
            actions.move_to_element(location_dropdown)
            actions.perform()
            
            wait.until(
            EC.presence_of_element_located((By.CLASS_NAME, "dropdown-content"))
            )
            categories_box=location_dropdown.find_element(By.CLASS_NAME, "dropdown-content")
            category_boxes=categories_box.find_elements(By.TAG_NAME, "a")
            category_boxes_links = [box.get_attribute("href") for box in category_boxes]
            categories = [box.get_attribute("textContent").strip() for box in category_boxes]
            
            print(categories)
            
            for category_link, CATEGORY_NAME in zip(category_boxes_links, categories):
                try:
                    time.sleep(random.uniform(2, 7))
                    #click on the category
                    driver.get(category_link)
                
                    # Get products
                    wait.until(EC.presence_of_all_elements_located(
                        (By.CLASS_NAME, "single_product_theme")
                    ))
                    infinite_scroll(driver)
                    product_cards = wait.until(EC.presence_of_all_elements_located(
                        (By.CLASS_NAME, "single_product_theme")
                    ))

                    for product_card in product_cards:
                        try:
                            product_link = product_card.find_element(By.TAG_NAME, "a").get_attribute("href")
                            name=product_card.find_element(By.CLASS_NAME, "product_name_theme").text.strip()
                            currency=product_card.find_element(By.CLASS_NAME, "item-currency").text.strip()
                            price=product_card.find_element(By.CLASS_NAME, "price-value").text.strip()
                            image_url=product_card.find_element(By.TAG_NAME, "img").get_attribute("src")
                            all_products.append({
                                "store": store_name,
                                "name": name,
                                "product-link": product_link,
                                "price": f"{currency} {price}",
                                "category":CATEGORY_NAME,
                                "image_url": image_url
                            })
                
                        except Exception as e:
                            print(f"[{store_name}] Error extracting product details: {str(e)}")
                            continue
                        
                    print (f"Category '{CATEGORY_NAME}': Found {len(product_cards)} products")

                except Exception as e:
                        print(f"[{store_name}] Error extracting category {str(e)}")
                        continue
    except Exception as e:
        print(f"An error occurred: {e}")
        driver.save_screenshot("error_screenshot.png")
        
    finally:
        return all_products

for imtiaz

def scrape_imtiaz_grocery_products(driver):
    try:
            store_name = "Imtiaz"
            IMTIAZ_GROCERY_URL = f"https://shop.imtiaz.com.pk/"
            
            driver.get(IMTIAZ_GROCERY_URL)
            wait = WebDriverWait(driver, 10)
            
            all_products = []
            # Select location
            try:
                area = wait.until(EC.presence_of_element_located(
                    (By.XPATH, "/html/body/div[2]/div[3]/div/div/div/div/div[3]/div[3]/div/div/input")
                ))
                area.send_keys(Keys.ENTER)
                area.send_keys(Keys.DOWN)
                area.send_keys(Keys.DOWN)
                area.send_keys(Keys.ENTER)
                
                submit_button = wait.until(EC.presence_of_element_located(
                    (By.XPATH, "/html/body/div[2]/div[3]/div/div/div/div/div[3]/button")
                ))
                submit_button.click()
            except TimeoutException:
                print(f"[{store_name}] No location box appeared")
            
            # Get products
            try:
                categories_tag=driver.find_element(By.XPATH, "/html/body/div[1]/div[2]/div/div[1]/div[3]/div[1]")
                categories_tag.click()
                categories_list=wait.until(EC.presence_of_element_located(
                    (By.XPATH, "/html/body/div[1]/div[2]/div/div[1]/div[3]/div[3]/ul")
                ))
                categories_list=categories_list.find_elements(By.TAG_NAME, "a")
                categpry_links=[]
                
                for i in categories_list:
                    categpry_links.append(i.get_attribute("href"))
                
                print(categpry_links)

                for link in categpry_links:
                    try:
                        
                        driver.get(link)
                        
                        sub_categories_container = wait.until(EC.element_to_be_clickable(
                            (By.XPATH, "/html/body/div[1]/div[2]/div/div[2]/div/div")
                        ))
                        
                        #print(sub_categories_container.text)
                        #break
                        
                        sub_categories = sub_categories_container.find_elements(By.TAG_NAME, "a")
                        sub_categories_links=[sub_category.get_attribute("href") for sub_category in sub_categories]
                        sub_categories_names=[sub_category.get_attribute("textContent").strip() for sub_category in sub_categories]
                        
                        #print(sub_categories_names)
                        #print(sub_categories_links)
                        #break
                        
                        for sub_category_link,sub_category_name in zip(sub_categories_links,sub_categories_names):
                            driver.get(sub_category_link)
                            products = []
                            current_url =None
                            #now gonna click on next button until it is disabled
                            while True:
                                try:
                                    wait.until(EC.url_changes(current_url))
                                    wait.until(EC.presence_of_all_elements_located(
                                        (By.CLASS_NAME, "hazle-product-item_product_item__FSm1N")
                                    ))
                                    infinite_scroll(driver)
                                    temp=wait.until(EC.presence_of_all_elements_located(
                                        (By.CLASS_NAME, "hazle-product-item_product_item__FSm1N")
                                    ))
                                    products= temp
                                    
                                    
                                    for product in products:
                                        try:
                                            product_text_container=product.find_element(By.CLASS_NAME, "hazle-product-item_product_item_text_container__Apuq1")
                                            
                                            product_name = product_text_container.find_element(By.CLASS_NAME, "hazle-product-item_product_item_description__ejRDa").text.strip()
                                            product_price = product_text_container.find_element(By.CLASS_NAME, "hazle-product-item_product_item_price_label__ET_we").text.strip()
                                            
                                            product_link_id = product.get_attribute("id")
                                            product_link = f"https://shop.imtiaz.com.pk/product/{product_link_id}"
                                            
                                            image_url = product.find_element(By.TAG_NAME, "img").get_attribute("src")
                                            all_products.append({
                                                    "store": store_name,
                                                    "name": product_name,
                                                    "product-link": product_link,
                                                    "price": product_price,
                                                    "category":sub_category_name,
                                                    "image_url": image_url
                                                })
                                        except Exception as e:
                                            print(f"[{store_name}] Error extracting product info: {str(e)}")
                                            continue
                                        
                                    print (f"Sub-Category '{sub_category_name}': Found {len(products)} products")
                                        
                                        
                                        
                                    button = driver.find_element(By.XPATH, "//button[normalize-space()='Next']")
                                        
                                    if button.get_attribute("disabled"):
                                        print("Next button is disabled")
                                        break
                                    else:
                                        current_url = driver.current_url
                                        button.click()
                                except NoSuchElementException:
                                        print("Next button not found, assuming last page")
                                        break
                                
                        
                    except Exception as e:
                                print(f"[{store_name}] Error extracting sub-category {str(e)}")
                                continue
                        
            except TimeoutException:
                print(f"[{store_name}] No products found")
                
            
    except Exception as e:
             print(f"An error occurred: {e}")
             driver.save_screenshot("error_screenshot.png")
            
    finally:
      return all_products

for metro

def scrape_metro_grocery_products(driver):
        try:
                store_name = "Metro"
                METRO_GROCERY_URL = f"https://www.metro-online.pk/"
                        
                categories_url="https://www.metro-online.pk/category/"

                driver.get(METRO_GROCERY_URL)
                
                time.sleep(random.uniform(6, 10))
                
                wait=WebDriverWait(driver, 10)
                
                three_bars_icon=wait.until(EC.presence_of_element_located((By.CLASS_NAME, "NewDesktopNav_menu_outlined_icon_container__8Lrtz")))
                time.sleep(random.uniform(6, 10))
                three_bars_icon.click()
                
                grocery_categories_container=wait.until(EC.presence_of_element_located((By.CLASS_NAME, "CategoryListingWeb_catergory_listing_Grocery_container__qM57p")))
                
                category_boxes=grocery_categories_container.find_elements(By.CLASS_NAME, "CategoryListingWeb_category_listing_container__wJJOm")
                
                text=[]
                for category_box in category_boxes:
                        #make an action chain to click on category box
                        time.sleep(random.uniform(2, 5))
                        actions=ActionChains(driver)
                        actions.move_to_element(category_box)
                        actions.perform()
                        
                        try:
                                elemens=wait.until(EC.presence_of_all_elements_located((By.CLASS_NAME, "CategoryListingWeb_category_expanded_item_level_two__XkUxr")))
                                for element in elemens:
                                        try:
                                                store_elements_array=None
                                                store_text=[]
                                                category_text=[]
                                                #####################################
                                                store_elements_array=WebDriverWait(element, 3).until(EC.presence_of_element_located((By.CLASS_NAME, "CategoryListingWeb_category_expanded_level_three_container__7eQeE")))
                
                                                store_elements_array=store_elements_array.find_elements(By.CLASS_NAME, "CategoryListingWeb_category_expanded_level_three_item__jjFUX")
                                                
                                                for store_element in store_elements_array:
                                                        store_text.append(store_element.text.strip().lower().replace(" ","-"))
                                                        
                                                #####################################       
                                                category_element=WebDriverWait(element, 3).until(EC.presence_of_element_located((By.CLASS_NAME, "CategoryListingWeb_category_expanded_item_level_two_title_container__Z5dTt"))) 
                                                category_text.append(category_element.text.strip().lower().replace(" ","-"))
                                
                                                ################################
                                                if len(store_text) > 0:
                                                        print(f"[Metro] Sub-categories found for {category_text[-1]}: {store_text}")
                                                        text.extend(store_text)
                                                else:
                                                        print(f"[Metro] No sub-categories found for {category_text[-1]}")
                                                        text.extend(category_text)
                                                
                                                
                                                
                                        except Exception as e:
                                                print(f"[Metro] Error extracting sub-categories for {category_text[-1]}: {str(e)}")
                                                continue

                        except Exception as e:
                                print(f"[Metro] Error extracting categories for {category_box.text}: {str(e)}")
                                continue

                products_details = []
                for category in text:
                        try:
                                time.sleep(random.uniform(2, 7))
                                category_url=f"https://www.metro-online.pk/store/{category}"
                                print(f"Navigating to category URL: {category_url}")
                                
                                
                                driver.get(category_url)
                                
                                wait.until(EC.presence_of_all_elements_located(
                                (By.CLASS_NAME, "CategoryGrid_product_card__FUMXW")
                                ))
                                
                                infinite_scroll(driver)
                        
                                product_cards = wait.until(EC.presence_of_all_elements_located( (By.CLASS_NAME, "CategoryGrid_product_card__FUMXW")
                                ))
                                
                                for product_card in product_cards:
                                        try:
                                                
                                                product_link=product_card.find_element(By.TAG_NAME, "a").get_attribute("href")
                                                name = product_card.find_element(By.CLASS_NAME, "CategoryGrid_product_name__3nYsN").text
                                                price = product_card.find_element(By.CLASS_NAME, "CategoryGrid_product_price__Svf8T").text
                                                image_container=product_card.find_element(By.CLASS_NAME, "CategoryGrid_productImg_container__Ga1ll")
                                                image_url=image_container.find_element(By.TAG_NAME, "img").get_attribute("src")
                                                products_details.append({
                                                        "store": store_name,
                                                        "name": name,
                                                        "product-link": product_link,
                                                        "price": price,
                                                        "category":category,
                                                        "image_url":image_url
                                                })
                                        except Exception as e:
                                                print(f"[{store_name}] Error extracting product details: {str(e)}")
                                                continue
                                
                                print (f"Category '{category}': Found {len(product_cards)} products")
                                print(products_details[-1])
                                
                        except Exception as e:
                                print(f"[Metro] Error in navigating to categories: {str(e)}")
                                driver.save_screenshot("error_screenshot.png")
                           
        except Exception as e:
                print(f"[Metro] Error during scraping: {str(e)}")
        finally:
                return products_details
        


al_fatah_products=scraope_al_fateh_grocery_products(driver)
df_fateh=pd.DataFrame(al_fatah_products)
df_fateh.to_csv("al-fateh-products.csv")


metro_products=scrape_metro_grocery_products(driver)
df_metro=pd.DataFrame(metro_products)
df_metro.to_csv("Metro-products.csv")


imtiaz_products=scrape_imtiaz_grocery_products(driver)
df_imtiaz=pd.DataFrame(imtiaz_products)
df_imtiaz.to_csv("Imtiaz-products.csv")


jalal_sons_products=scrape_jalal_sons_grocery_products(driver)
df_jalalsons=pd.DataFrame(jalal_sons_products)
df_jalalsons.to_csv("jalal-sons-products.csv")
