==> 3)creating_canonicalized_categories.ipynb <==
import pandas as pd
import re

df_fateh=pd.read_csv("al-fateh-products.csv")
df_imtiaz=pd.read_csv("Imtiaz-products.csv")
df_jalalsons=pd.read_csv("jalal-sons-products.csv")
df_metro=pd.read_csv("Metro-products.csv")

we are gonna create cleaned_catgories for all stores (canonicalization of categories)

for fateh

print("shape of fateh data frame:")
print(df_fateh.shape)


print("Sample data from fateh data frame:")
display(df_fateh.sample(10))

print("categories in fateh data frame:")
fateh_categories=df_fateh['category'].value_counts()
print(fateh_categories)

pattern=r' +'
replacement=r'_'
fateh_categories_json=dict.fromkeys([ re.sub(pattern, replacement, c.lower()) + "---fateh" for c in fateh_categories.index])
print(fateh_categories_json)



for imtiaz

print("shape of imtiaz data frame:")
print(df_imtiaz.shape)
print("#"*50)

print("Sample data from imtiaz data frame:")
display(df_imtiaz.sample(10))
print("#"*50)

print("categories in imtiaz data frame:")
imtiaz_categories=df_imtiaz['category'].value_counts()
print(imtiaz_categories)

pattern=r' +'
replacement=r'_'
imtiaz_categories_json=dict.fromkeys([ re.sub(pattern, replacement, c.lower()) + "---imtiaz" for c in imtiaz_categories.index])
print(imtiaz_categories_json)

for jalalsons

print("shape of jalalsons data frame:")
print(df_jalalsons.shape)
print("#"*50)

print("Sample data from jalalsons data frame:")
display(df_jalalsons.sample(10))
print("#"*50)

print("categories in jalalsons data frame:")
jalalsons_categories=df_jalalsons['category'].value_counts()
print(jalalsons_categories)

pattern=r' +'
replacement=r'_'
jalalsons_categories_json=dict.fromkeys([ re.sub(pattern, replacement, c.lower()) + "---jalalsons" for c in jalalsons_categories.index])
print(jalalsons_categories_json)

for metro

print("shape of metro data frame:")
print(df_metro.shape)
print("#"*50)

print("Sample data from metro data frame:")
display(df_metro.sample(10))
print("#"*50)

print("categories in metro data frame:")
metro_categories=df_metro['category'].value_counts()
print(metro_categories[metro_categories>10])  # to handle NaN values

pattern=r'-+'
replacement=r'_'
metro_categories_json=dict.fromkeys([ re.sub(pattern, replacement, c.lower()) + "---metro" for c in metro_categories.index])
print(metro_categories_json)

print("Total categories across all shops:")
print(len(metro_categories_json)+len(fateh_categories_json)+len(imtiaz_categories_json)+len(jalalsons_categories_json))





mapping categories across the stores to a common category

category_mapping = {
    'air_fresheners_&_home_fragrances---imtiaz': 'Home Fragrances',
    'air_sprays---metro': 'Home Fragrances',
    'anti_perspirants---metro': 'Personal Care',
    'antibacterial---metro': 'Personal Care',
    'baby_cereals---metro': 'Baby Food',
    'baby_feeding_accessories---imtiaz': 'Baby Care',
    'baby_food---fateh': 'Baby Food',
    'baby_food---imtiaz': 'Baby Food',
    'baby_milk---metro': 'Baby Food',
    'baby_milk_and_food---jalalsons': 'Baby Food',
    'baking_accessories---fateh': 'Baking Supplies',
    'baking_chocolates---fateh': 'Baking Supplies',
    'baking_goods---jalalsons': 'Baking Supplies',
    'baking_items---fateh': 'Baking Supplies',
    'beef---metro': 'Meat',
    'beverages---jalalsons': 'Beverages',
    'bins_and_buckets---metro': 'Home & Kitchen',
    'biscuits---fateh': 'Snacks',
    'biscuits_&_wafers---imtiaz': 'Snacks',
    'biscuits_,_crisps_and_snacks---jalalsons': 'Snacks',
    'biscuits_and_wafers---metro': 'Snacks',
    'body_sprays_and_body_mists---metro': 'Personal Care',
    'bread---fateh': 'Bakery',
    'breads---metro': 'Bakery',
    'buns---metro': 'Bakery',
    'butter---fateh': 'Dairy',
    'butter---jalalsons': 'Dairy',
    'butter---metro': 'Dairy',
    'cake_mixes_and_baking_add_ons---metro': 'Baking Supplies',
    'cakes_and_chocolates---metro': 'Confectionery',
    'candies_&_bubble_gums---fateh': 'Confectionery',
    'canned/bottled_foods---imtiaz': 'Canned Foods',
    'canned_foods_and_milks---jalalsons': 'Canned Foods',
    'canned_fruits---metro': 'Canned Foods',
    'canned_vegetables---metro': 'Canned Foods',
    'carbonated_soft_drinks---imtiaz': 'Beverages',
    'cereal,_jams_and_spreads---jalalsons': 'Breakfast Foods',
    'cereals---fateh': 'Breakfast Foods',
    'cereals---metro': 'Breakfast Foods',
    'cereals_&_oats---imtiaz': 'Breakfast Foods',
    'cheese---fateh': 'Dairy',
    'cheese---metro': 'Dairy',
    'chicken---metro': 'Meat',
    'chips_&_nimko---fateh': 'Snacks',
    'chocolates---fateh': 'Confectionery',
    'cleaning---imtiaz': 'Cleaning Supplies',
    'cleaning_products---jalalsons': 'Cleaning Supplies',
    'coffee---fateh': 'Beverages',
    'coffee---imtiaz': 'Beverages',
    'coffee_and_whiteners---metro': 'Beverages',
    'conditioners---metro': 'Personal Care',
    'confectionery_and_chocolates---jalalsons': 'Confectionery',
    'cookies---metro': 'Snacks',
    'cosmetics---jalalsons': 'Personal Care',
    'cream---metro': 'Dairy',
    'creams---metro': 'Dairy',
    'crisps_and_popcorn---metro': 'Snacks',
    'crockery---metro': 'Home & Kitchen',
    'cutlery---metro': 'Home & Kitchen',
    'dairy_creams---fateh': 'Dairy',
    'dessert---imtiaz': 'Desserts',
    'detergents_and_laundry_soaps---metro': 'Cleaning Supplies',
    'diapers_&_pants---imtiaz': 'Baby Care',
    'diapers_and_pampers---jalalsons': 'Baby Care',
    'diapers_and_wipes---metro': 'Baby Care',
    'dishwashing_bars---metro': 'Cleaning Supplies',
    'dishwashing_liquids---metro': 'Cleaning Supplies',
    'disinfectants---metro': 'Cleaning Supplies',
    'disposable---imtiaz': 'Disposables',
    'disposables---metro': 'Disposables',
    'drinking_powders---fateh': 'Beverages',
    'drinking_water---fateh': 'Beverages',
    'dry_fruits---imtiaz': 'Dry Fruits & Nuts',
    'dry_fruits_&_dates---fateh': 'Dry Fruits & Nuts',
    'dry_fruits_and_dates---jalalsons': 'Dry Fruits & Nuts',
    'dry_fruits_and_nuts---metro': 'Dry Fruits & Nuts',
    'edible_oil_&_ghee---imtiaz': 'Cooking Oils',
    'eggs---fateh': 'Dairy & Eggs',
    'fabric_care---metro': 'Cleaning Supplies',
    'facewashes---metro': 'Personal Care',
    'fine_life---metro': 'Store Brand',
    'fish---metro': 'Seafood',
    'flavored_milk---metro': 'Dairy',
    'flavoured_milk---fateh': 'Dairy',
    'flavoured_milk---imtiaz': 'Dairy',
    'flour,_rice_and_pulses---jalalsons': 'Staples',
    'flour---fateh': 'Staples',
    'flour---imtiaz': 'Staples',
    'flour---metro': 'Staples',
    'fragrant---metro': 'Home Fragrances',
    'fresh_fruits---metro': 'Fresh Produce',
    'fresh_milk_and_eggs---jalalsons': 'Dairy & Eggs',
    'fresh_vegetables---metro': 'Fresh Produce',
    'frozen_foods---jalalsons': 'Frozen Foods',
    'frozen_fries---fateh': 'Frozen Foods',
    'frozen_fries---metro': 'Frozen Foods',
    'frozen_items---fateh': 'Frozen Foods',
    'frozen_meat---metro': 'Frozen Foods',
    'frozen_mixed_fruits_and_vegetables---metro': 'Frozen Foods',
    'frozen_seafood---metro': 'Frozen Foods',
    'ghee---fateh': 'Cooking Oils',
    'ghee_and_oil---jalalsons': 'Cooking Oils',
    'glassware---metro': 'Home & Kitchen',
    'hair_colors---metro': 'Personal Care',
    'hand_and_body_washes---metro': 'Personal Care',
    'hangers_and_accessories---metro': 'Home & Kitchen',
    'honey---fateh': 'Spreads',
    'hygiene---imtiaz': 'Personal Care',
    'ice_cream---fateh': 'Frozen Foods',
    'ice_cream---metro': 'Frozen Foods',
    'iced_tea_and_coffee---metro': 'Beverages',
    'imported_drinks_&_juices---fateh': 'Beverages',
    'insecticides---metro': 'Home Care',
    'instant_drinks---metro': 'Beverages',
    'instant_tea_&_coffee---imtiaz': 'Beverages',
    'jams,_honey_and_spreads---metro': 'Spreads',
    'jams---fateh': 'Spreads',
    'jellies_and_custards---metro': 'Desserts',
    'juices---metro': 'Beverages',
    'juices_&_nectars---imtiaz': 'Beverages',
    'kebab_and_koftas---metro': 'Meat',
    'ketchup,_sauce_and_mayo---jalalsons': 'Condiments',
    'kitchen_utensils_and_accessories---metro': 'Home & Kitchen',
    'laundry---imtiaz': 'Cleaning Supplies',
    'liquid_tin_milk---fateh': 'Dairy',
    'local_drinks---fateh': 'Beverages',
    'lotions_and_sunscreen---metro': 'Personal Care',
    'make_to_drink---imtiaz': 'Beverages',
    'margarine---fateh': 'Dairy',
    'margarine---metro': 'Dairy',
    'mayo_&_spreads---fateh': 'Condiments',
    'men_grooming---metro': 'Personal Care',
    'metro_chef---metro': 'Store Brand',
    'metro_post_grocery---metro': 'Store Brand',
    'metro_professionals---metro': 'Store Brand',
    'milk---fateh': 'Dairy',
    'milk---imtiaz': 'Dairy',
    'milk_powder_&_whitener---fateh': 'Dairy',
    'mineral_water---jalalsons': 'Beverages',
    'mops_and_brooms---metro': 'Cleaning Supplies',
    'mutton---metro': 'Meat',
    'noodles_&_pasta---fateh': 'Pasta & Noodles',
    'noodles_&_pasta---imtiaz': 'Pasta & Noodles',
    'noodles_and_pasta---metro': 'Pasta & Noodles',
    'nuggets_and_snacks---metro': 'Frozen Foods',
    'nutritional_drinks---metro': 'Beverages',
    'o.t.c_medicines---metro': 'Health & Wellness',
    'office_supplies---metro': 'Home & Kitchen',
    'oil_and_ghee---metro': 'Cooking Oils',
    'oils---fateh': 'Cooking Oils',
    'oils_and_serums---metro': 'Personal Care',
    'ok---metro': 'Store Brand',
    'olive_oil---fateh': 'Cooking Oils',
    'oral_care---imtiaz': 'Personal Care',
    'other_food_items---jalalsons': 'Miscellaneous',
    'parathas---metro': 'Frozen Foods',
    'personal_hygiene---jalalsons': 'Personal Care',
    'pet_food---jalalsons': 'Pet Care',
    'pickle_&_vinegar---fateh': 'Condiments',
    'pickles_and_olives---metro': 'Condiments',
    'plain_eggs---metro': 'Dairy & Eggs',
    'plain_yogurt_and_flavored_yogurt---metro': 'Dairy',
    'popcorn---fateh': 'Snacks',
    'pots_and_pans---metro': 'Home & Kitchen',
    'powder_milk---jalalsons': 'Dairy',
    'powdered_milk---metro': 'Dairy',
    'prawns_and_crabs---metro': 'Seafood',
    'pulses---metro': 'Staples',
    'pulses_&_grains---imtiaz': 'Staples',
    'raita---fateh': 'Dairy',
    'red_syrups---metro': 'Beverages',
    'rice---imtiaz': 'Staples',
    'rice---metro': 'Staples',
    'rice_products---fateh': 'Staples',
    'rusks---metro': 'Bakery',
    'rusks_&_buns---fateh': 'Bakery',
    'salt,_spices_&_herbs---imtiaz': 'Spices & Seasonings',
    'salt---fateh': 'Spices & Seasonings',
    'samosas_and_rolls---metro': 'Frozen Foods',
    'sauces,_dressings_&_seasonings---imtiaz': 'Condiments',
    'sauces_&_soups---fateh': 'Condiments',
    'sauces_and_seasonings---metro': 'Condiments',
    'sausages---metro': 'Meat',
    'school_essentials---metro': 'Home & Kitchen',
    'shampoos---metro': 'Personal Care',
    'shoe_polish---metro': 'Home Care',
    'skin_care---imtiaz': 'Personal Care',
    'soft_drinks---metro': 'Beverages',
    'spices_and_herbs---metro': 'Spices & Seasonings',
    'spices_and_miscellaneous---jalalsons': 'Spices & Seasonings',
    'sponges---metro': 'Cleaning Supplies',
    'sports_drink---imtiaz': 'Beverages',
    'spreads---imtiaz': 'Spreads',
    'squashes---fateh': 'Beverages',
    'squashes---metro': 'Beverages',
    'storage_containers---metro': 'Home & Kitchen',
    'sugar---fateh': 'Staples',
    'sugar---imtiaz': 'Staples',
    'sugar---metro': 'Staples',
    'sweets_and_toffees---metro': 'Confectionery',
    'tea---imtiaz': 'Beverages',
    'tea---metro': 'Beverages',
    'tea_and_coffee---jalalsons': 'Beverages',
    'tea_whiteners---imtiaz': 'Beverages',
    'teas---fateh': 'Beverages',
    'tin_foods---fateh': 'Canned Foods',
    'tissue---imtiaz': 'Disposables',
    'tissues_and_napkins---metro': 'Disposables',
    'tissues_and_sanitary---jalalsons': 'Disposables',
    'toilet_supplies---metro': 'Personal Care',
    'toiletries---jalalsons': 'Personal Care',
    'tooth_brushes---metro': 'Personal Care',
    'tooth_pastes---metro': 'Personal Care',
    'traditional_mixes---metro': 'Cooking Ingredients',
    'uht_and_pasteurized_milk---metro': 'Dairy',
    'water---imtiaz': 'Beverages',
    'water---metro': 'Beverages',
    'women_care---metro': 'Personal Care',
    'wraps_and_pitta---metro': 'Bakery',
    'yoghurt,_butter,_cream_and_cheese---jalalsons': 'Dairy',
    'yogurt---fateh': 'Dairy'
}

# Print summary of common categories
common_categories = set(category_mapping.values())
print(f"Total store-specific categories: {len(category_mapping)}")
print(f"Total common categories: {len(common_categories)}")
print(f"\nCommon categories: {sorted(common_categories)}")

df_fateh.category=df_fateh.category.str.lower()
df_fateh.category=[re.sub(r' +',r'_',c)+"---fateh" for c in df_fateh.category]

df_imtiaz.category=df_imtiaz.category.str.lower()
df_imtiaz.category=[re.sub(r' +',r'_',c)+"---imtiaz" for c in df_imtiaz.category]

df_jalalsons.category=df_jalalsons.category.str.lower()
df_jalalsons.category=[re.sub(r' +',r'_',c)+"---jalalsons" for c in df_jalalsons.category]

df_metro.category=df_metro.category.str.lower()
df_metro.category=[re.sub(r'-+',r'_',c)+"---metro" for c in df_metro.category]


df_fateh['common_category']=df_fateh['category'].map(category_mapping)
df_imtiaz['common_category']=df_imtiaz['category'].map(category_mapping)
df_jalalsons['common_category']=df_jalalsons['category'].map(category_mapping)
df_metro['common_category']=df_metro['category'].map(category_mapping)  

print(df_fateh[['category','common_category']].sample(20))
print(df_imtiaz[['category','common_category']].sample(20))
print(df_jalalsons[['category','common_category']].sample(20))
print(df_metro[['category','common_category']].sample(20))

#df_fateh.to_csv("al-fateh-products-with-common-categories.csv", index=False)
#df_imtiaz.to_csv("Imtiaz-products-with-common-categories.csv", index=False)
#df_jalalsons.to_csv("jalal-sons-products-with-common-categories.csv", index=False)
#df_metro.to_csv("metro-products-with-common-categories.csv", index=False)

print("Common category value counts:")
print("----------------------------")
print("For df_fateh:")
print(df_fateh["common_category"].value_counts())
print("For df_imtiaz:")
print(df_imtiaz["common_category"].value_counts())
print("For df_jalalsons:")
print(df_jalalsons["common_category"].value_counts())
print("For df_metro:")
print(df_metro["common_category"].value_counts())

selecting the common categories across all the stores for model training

common_categories_from_all_stores = set(df_fateh['common_category']).intersection(set(df_imtiaz['common_category'])).intersection(set(df_jalalsons['common_category'])).intersection(set(df_metro['common_category']))

all_common_categories = set(df_fateh['common_category']).union(set(df_imtiaz['common_category'])).union(set(df_jalalsons['common_category'])).union(set(df_metro['common_category']))

print("Number of common categories across all stores:")
print(len(common_categories_from_all_stores))
print("Number of all categories across the stores:")
print(len(all_common_categories))

print("Common categories across all stores:")
print(common_categories_from_all_stores)
print("Categories not present in all stores:")
print(all_common_categories.difference(common_categories_from_all_stores))

common_fateh=df_fateh[df_fateh['common_category'].isin(common_categories_from_all_stores)]
a=common_fateh.groupby('common_category').agg(length=("name","count"))

b=df_fateh[df_fateh['common_category'].isin(all_common_categories.difference(common_categories_from_all_stores))].groupby('common_category').agg(length=("name","count"))

print("Categories present in all stores:")
print(a)
print("Categories not present in all stores:")
print(b)

common_jalalsons=df_jalalsons[df_jalalsons['common_category'].isin(common_categories_from_all_stores)]
a=common_jalalsons.groupby('common_category').agg(length=("name","count"))

b=df_jalalsons[df_jalalsons['common_category'].isin(all_common_categories.difference(common_categories_from_all_stores))].groupby('common_category').agg(length=("name","count"))

print("Categories present in all stores:")
print(a)
print("Categories not present in all stores:")
print(b)

common_metro=df_metro[df_metro['common_category'].isin(common_categories_from_all_stores)]
a=common_metro.groupby('common_category').agg(length=("name","count"))
b=df_metro[df_metro['common_category'].isin(all_common_categories.difference(common_categories_from_all_stores))].groupby('common_category').agg(length=("name","count"))

print("Categories present in all stores:")
print(a)
print("Categories not present in all stores:")
print(b)

    
common_imtiaz=df_imtiaz[df_imtiaz['common_category'].isin(common_categories_from_all_stores)]
a=common_imtiaz.groupby('common_category').agg(length=("name","count"))

b=df_imtiaz[df_imtiaz['common_category'].isin(all_common_categories.difference(common_categories_from_all_stores))].groupby('common_category').agg(length=("name","count"))

print("Categories present in all stores:")
print(a)
print("Categories not present in all stores:")
print(b)



df=pd.concat([common_fateh, common_jalalsons, common_metro, common_imtiaz],ignore_index=True)
df.to_csv("grocery_products_canonicalized_categories.csv",index=False)
